Cycle over a known set of values.
