+++ src/stream/mod.rs
@@ -300,6 +300,7 @@
 //! [`take`]: trait.Stream.html#method.take
 //! [`min`]: trait.Stream.html#method.min
 
+pub use cycle::{cycle, Cycle};
 pub use empty::{empty, Empty};
 pub use from_fn::{from_fn, FromFn};
 pub use from_iter::{from_iter, FromIter};
@@ -312,6 +313,7 @@ pub use stream::{
 
 pub(crate) mod stream;
 
+mod cycle;
 mod empty;
 mod from_fn;
 mod from_iter;
